Refactored architecture for modular structure
