+++ auth.py
@@ -1,0 +1,8 @@
+# auth.py - Authentication and session management module
+
+users = {
+    'user': 'pass'
+}
+
+def authenticate(username, password):
+    return username in users and users[username] == password
