+++ conversions.py
@@ -1,0 +1,7 @@
+# conversions.py - Unit conversions module
+
+def celsius_to_fahrenheit(c):
+    return (c * 9/5) + 32
+
+def fahrenheit_to_celsius(f):
+    return (f - 32) * 5/9
