+++ memory.py
@@ -1,0 +1,9 @@
+# memory.py - Memory and history management module
+
+history = []
+
+def add_history(result):
+    history.append(result)
+
+def clear_history():
+    history.clear()
